Fullført hele sulamitten

+++ modules/db.js
@@ -23,10 +23,37 @@ dbMethods.createBlogPosts = function(heading, blogtext, userid){
     return pool.query(sql, values); //return the promise
 }
 //-----------------------------------------------
-dbMethods.deleteBlogPosts = function(id){
-    let sql = "DELETE FROM blogposts WHERE id = $1 RETURNING *";
-    let values = [id];
+dbMethods.deleteBlogPosts = function(id, userid){
+    let sql = "DELETE FROM blogposts WHERE id = $1 AND userid = $2 RETURNING *";
+    let values = [id, userid];
     return pool.query(sql, values);//return the promise
 }
+
+//--------
+dbMethods.getAllUsers = function() {
+    let sql = "SELECT id, username, password, salt FROM users";
+    return pool.query(sql); //return the promise
+}
+//--------
+dbMethods.getUser = function(username) {
+    let sql = "SELECT * FROM users WHERE username = $1";
+    let values = [username]; 
+    return pool.query(sql, values);
+}
+//--------
+dbMethods.createUser = function(username, password, salt) {
+    let sql = "INSERT INTO users (id, username, password, salt) VALUES(DEFAULT, $1, $2, $3) returning *";
+    let values = [username, password, salt];
+    return pool.query(sql, values);
+}
+//--------
+dbMethods.deleteUser = function(id) {
+    let sql = "DELETE FROM users WHERE id = $1 RETURNING *";
+    let values = [id];
+    return pool.query(sql, values);
+}
+//--------
+
+//--------
 // export dbMethods---------------------------------
 module.exports = dbMethods;
